Move some functions to higher-level module
